pyproject.toml
[project]
name = "agentcore-telco"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.3",
]
